requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
communex==0.1.32
cytoolz==0.12.3
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==4.3.1
eth-utils==2.2.2
exceptiongroup==1.2.1
fastapi==0.110.0
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
joblib==1.4.2
keras==3.4.1
keylimiter==0.1.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.3.0
multidict==6.0.5
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
protobuf==4.25.3
ptyprocess==0.7.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
scalecodec==1.2.10
scikit-learn==1.5.0
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
substrate-interface==1.7.10
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==0.12.1
typeguard==4.3.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.29.0
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
